refactor: import updated Select component
